Rename: change default components directory name
- add components directory in ui
- move default components to ui/components except Header.tsx
